fix(projects): move ProjectTestMixin to __init__
